backend/requirements.txt
fastapi
uvicorn
jinja2
aiofiles
pypandoc
spacy
nltk
transformers
python-multipart
sqlalchemy asyncpg 
databases
python-dotenv
